+++ packages/author-profile/package.json
@@ -58,10 +58,7 @@
     "@times-components/author-head": "0.2.0",
     "@times-components/card": "0.3.0",
     "@times-components/pagination": "0.2.0",
-    "@times-components/provider": "0.0.1",
-    "graphql-tag": "2.4.2",
     "jest-cli": "20.0.4",
-    "lodash.get": "4.4.2",
     "lodash.merge": "4.6.0",
     "prop-types": "15.5.10",
     "react": "16.0.0-alpha.6",
