+++ packages/provider/package.json
@@ -50,6 +50,9 @@
     "webpack": "2.6.1"
   },
   "dependencies": {
+    "@times-components/author-profile": "0.1.0",
+    "graphql-tag": "2.4.2",
+    "lodash.get": "4.4.2",
     "lodash.pick": "4.4.0",
     "prop-types": "15.5.10",
     "react": "16.0.0-alpha.6",
