+++ packages/provider/provider.stories.js
@@ -1,9 +1,11 @@
 import React from "react";
 import { Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
-import { gql } from "react-apollo";
+import { ApolloProvider, gql } from "react-apollo";
 import { withApolloProvider } from "storybook-addon-apollo-graphql";
-import connectGraphql from "./provider.js";
+
+import client from "./apollo-client";
+import connectGraphql, { AuthorProfileProvider } from "./provider.js";
 
 const Component = props => <Text>{JSON.stringify(props, null, 2)}</Text>;
 
@@ -25,3 +27,14 @@ storiesOf("Provider", module)
     const WithData = connectGraphql(query)(Component);
     return <WithData />;
   });
+
+storiesOf("Provider", module).add("AuthorProfileProvider", () =>
+  <ApolloProvider client={client}>
+    <AuthorProfileProvider
+      slug="fiona-hamilton"
+      page={1}
+      pageSize={10}
+      imageRatio="3:2"
+    />
+  </ApolloProvider>
+);
